Улучшение читаемости кода программы.
