+++ src/test/java/ru/yandex/practicum/filmorate/UserControllerTests.java
@@ -1,0 +1,70 @@
+package ru.yandex.practicum.filmorate;
+
+import jakarta.validation.ConstraintViolation;
+import jakarta.validation.Validation;
+import jakarta.validation.Validator;
+import jakarta.validation.ValidatorFactory;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.boot.test.web.client.TestRestTemplate;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import ru.yandex.practicum.filmorate.model.User;
+
+import java.time.LocalDate;
+import java.util.Set;
+
+import static org.junit.jupiter.api.Assertions.*;
+
+@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.DEFINED_PORT)
+
+public class UserControllerTests {
+    @Autowired
+    private TestRestTemplate restTemplate;
+
+    private Validator validator;
+
+    @BeforeEach
+    void setUp() {
+        ValidatorFactory factory = Validation.buildDefaultValidatorFactory();
+        validator = factory.getValidator();
+    }
+
+    @Test
+    void userValidData() {
+        User user = createValidUser();
+        Set<ConstraintViolation<User>> violations = validator.validate(user);
+        assertTrue(violations.isEmpty());
+    }
+
+    @Test
+    void userInvalidEmail() {
+        User user = createValidUser();
+        user.setEmail("invalid-email");
+        assertSingleViolation(user, "Некорректный формат email.");
+    }
+
+    @Test
+    void userCreateValidUser() {
+        User user = createValidUser();
+        ResponseEntity<User> response = restTemplate.postForEntity("/users", user, User.class);
+        assertEquals(HttpStatus.OK, response.getStatusCode());
+        assertNotNull(response.getBody().getId());
+    }
+
+    private User createValidUser() {
+        User user = new User();
+        user.setEmail("valid@test.com");
+        user.setLogin("validLogin");
+        user.setBirthday(LocalDate.of(2000, 1, 1));
+        return user;
+    }
+
+    private <T> void assertSingleViolation(T object, String expectedMessage) {
+        Set<ConstraintViolation<T>> violations = validator.validate(object);
+        assertEquals(1, violations.size());
+        assertEquals(expectedMessage, violations.iterator().next().getMessage());
+    }
+}
